// src/meta/common/interfaces.ts
export interface VideoParams {
    /** Whether to include the description in the response. */
    description?: boolean;
    /** Whether to include ad breaks in the response. */
    adBreaks?: boolean;
    /** Whether to include back-dated time in the response. */
    backDatedTime?: boolean;
    /** Whether to include content category in the response. */
    contentCategory?: boolean;
    /** Whether to include content tags in the response. */
    contentTags?: boolean;
    /** Limit the number of items returned. */
    limit?: number;
    /** Whether to include back-dated time granularity in the response. */
    backDatedTimeGranularity?: boolean;
    /** Whether to include custom labels in the response. */
    customLabels?: boolean;
    /** Whether to include created time in the response. */
    createdTime?: boolean;
    /** Whether to include embed HTML in the response. */
    embedHtml?: boolean;
    /** Whether to include embeddable information in the response. */
    embeddable?: boolean;
    /** Whether to include event information in the response. */
    event?: boolean;
    /** Whether to include format information in the response. */
    formate?: boolean;
    /** Whether to include 'from' information in the response. */
    from?: boolean;
    /** Whether to include icon information in the response. */
    icon?: boolean;
    /** Whether to include ID information in the response. */
    id?: boolean;
    /** Whether to include crosspost video information in the response. */
    isCrosspostVideo?: boolean;
    /** Whether to include episode information in the response. */
    isEpisode?: boolean;
    /** Whether to include crossposting eligibility information in the response. */
    isCrosspostingEligible?: boolean;
    /** Whether to include Instagram eligibility information in the response. */
    isInstagramEligible?: boolean;
    /** Whether to include length information in the response. */
    length?: boolean;
    /** Whether to include live status information in the response. */
    liveStatus?: boolean;
    /** Whether to include thumbnails in the response. */
    thumbnails?: boolean;
    /** Whether to include poll settings in the response. */
    pollSettings?: boolean;
    /** Whether to include place information in the response. */
    place?: boolean;
    /** Whether to include post ID information in the response. */
    postId?: boolean;
    /** Whether to include post views information in the response. */
    postViews?: boolean;
    /** Whether to include privacy information in the response. */
    privacy?: boolean;
    /** Whether to include source information in the response. */
    source?: boolean;
    /** Whether to include published information in the response. */
    published?: boolean;
    /** Whether to include status information in the response. */
    status?: boolean;
    /** Whether to include title information in the response. */
    title?: boolean;
    /** Whether to include universal video ID information in the response. */
    universalVideoId?: boolean;
    /** Whether to include updated time information in the response. */
    updatedTime?: boolean;
    /** Whether to include views information in the response. */
    views?: boolean;
    /** Whether to include captions in the response. */
    captions?: boolean;
    /** Whether to include polls in the response. */
    polls?: boolean;
    /** Whether to include picture information in the response. */
    picture?: boolean;
    /** Whether to include permalink URL in the response. */
    permaLinkUrl?: boolean;
    /** Whether to include likes information in the response. */
    likes?: boolean;
    /** Whether to include tags in the response. */
    tags?: boolean;
    /** Whether to include sponsor tags in the response. */
    sponsorTags?: boolean;
    /** Whether to include collaborators information in the response. */
    collaborators?: boolean;
    /** Whether to include scheduled publish time information in the response. */
    scheduledPublishTime?: boolean;
    /** Whether to include comments in the response. */
    comments?: boolean;
    /** Whether to include premiere living rooms status information in the response. */
    premiereLivingRoomsStatus?: boolean;
    /** Whether to include crosspost shared pages information in the response. */
    crosspostSharedPages?: boolean;
    /** Whether to include video insights in the response. */
    videoInsights?: boolean;
}

/**
 * Interface for parameters related to photo data.
 */
export interface PhotoParams {
    /** Limit the number of items returned. */
    limit?: number;
    /** Whether to include album information in the response. */
    album?: boolean;
    /** Whether to include images in the response. */
    images?: boolean;
    /** Whether to include icon information in the response. */
    icon?: boolean;
    /** Whether to include alt text in the response. */
    altText?: boolean;
    /** Whether to include back-dated time in the response. */
    backdatedTime?: boolean;
    /** Whether to include custom alt text in the response. */
    altTextCustom?: boolean;
    /** Whether to include back-dated time granularity in the response. */
    backDatedTimeGranularity?: boolean;
    /** Whether to include back-dated capability in the response. */
    canBackdated?: boolean;
    /** Whether to include delete capability in the response. */
    canDelete?: boolean;
    /** Whether to include tag capability in the response. */
    canTag?: boolean;
    /** Whether to include link information in the response. */
    link?: boolean;
    /** Whether to include name information in the response. */
    name?: boolean;
    /** Whether to include name tags in the response. */
    nameTags?: boolean;
    /** Whether to include page story ID in the response. */
    pageStoryId?: boolean;
    /** Whether to include webp images in the response. */
    webpImages?: boolean;
    /** Whether to include picture information in the response. */
    picture?: boolean;
    /** Whether to include created time in the response. */
    createdTime?: boolean;
    /** Whether to include events in the response. */
    events?: boolean;
    /** Whether to include 'from' information in the response. */
    from?: boolean;
    /** Whether to include height information in the response. */
    height?: boolean;
    /** Whether to include width information in the response. */
    width?: boolean;
    /** Whether to include ID information in the response. */
    id?: boolean;
    /** Whether to include place information in the response. */
    place?: boolean;
    /** Whether to include target information in the response. */
    target?: boolean;
    /** Whether to include updated time in the response. */
    updatedTime?: boolean;
    /** Whether to include comments in the response. */
    comments?: boolean;
    /** Whether to include likes in the response. */
    likes?: boolean;
    /** Whether to include sponsor tags in the response. */
    sponsorsTag?: boolean;
    /** Whether to include create time in the response. */
    createTime?: boolean;
    /** Whether to include cover information in the response. */
    cover?: boolean;
}
// ...
export interface EventsParams {
    /** Limit the number of items returned. */
    limit?: number;
    /** Whether to include attending count in the response. */
    attendingCount?: boolean;
    /** Whether to include guest invite capability in the response. */
    canGuestInvite?: boolean;
    /** Whether to include category information in the response. */
    category?: boolean;
    /** Whether to include cover information in the response. */
    cover?: boolean;
    /** Whether to include declined count in the response. */
    declinedCount?: boolean;
    /** Whether to include created time in the response. */
    createdTime?: boolean;
    /** Whether to include discount code enabled information in the response. */
    discountCodeEnabled?: boolean;
    /** Whether to include end time in the response. */
    endTime?: boolean;
    /** Whether to include description in the response. */
    description?: boolean;
    /** Whether to include guest list enabled information in the response. */
    guestListEnabled?: boolean;
    /** Whether to include event times in the response. */
    eventTimes?: boolean;
    /** Whether to include ID information in the response. */
    id?: boolean;
    /** Whether to include canceled status in the response. */
    isCanceled?: boolean;
    /** Whether to include interested count in the response. */
    intrestedCount?: boolean;
    /** Whether to include draft status in the response. */
    isDraft?: boolean;
    /** Whether to include online status in the response. */
    isOnline?: boolean;
    /** Whether to include page ownership information in the response. */
    isPageOwned?: boolean;
    /** Whether to include maybe count in the response. */
    maybeCount?: boolean;
    /** Whether to include name information in the response. */
    name?: boolean;
    /** Whether to include online event format in the response. */
    onlineEventFormate?: boolean;
    /** Whether to include online event third-party URL in the response. */
    onlineEventThirdPartyUrl?: boolean;
    /** Whether to include owner information in the response. */
    owner?: boolean;
    /** Whether to include place information in the response. */
    place?: boolean;
    /** Whether to include scheduled publish time in the response. */
    shcheduledPublishTime?: boolean;
    /** Whether to include start time in the response. */
    startTime?: boolean;
    /** Whether to include ticket URI in the response. */
    ticketUri?: boolean;
    /** Whether to include ticket URI start sales time in the response. */
    ticketUriStartSalesTime?: boolean;
    /** Whether to include ticketing privacy URI in the response. */
    ticketingPrivacyUri?: boolean;
    /** Whether to include ticketing term URI in the response. */
    ticketingTermUri?: boolean;
    /** Whether to include timezone information in the response. */
    timezone?: boolean;
    /** Whether to include type information in the response. */
    type?: boolean;
    /** Whether to include updated time in the response. */
    updatedTime?: boolean;
    /** Whether to include comments in the response. */
    comments?: boolean;
    /** Whether to include no reply count in the response. */
    noreplyCount?: boolean;
    /** Whether to include feed information in the response. */
    feed?: boolean;
    /** Whether to include photos in the response. */
    photos?: boolean;
    /** Whether to include picture information in the response. */
    picture?: boolean;
    /** Whether to include live videos in the response. */
    liveVidoes?: boolean;
    /** Whether to include roles in the response. */
    roles?: boolean;
    /** Whether to include ticket tiers in the response. */
    ticketTiers?: boolean;
    /** Whether to include videos in the response. */
    Videos?: boolean;
}